Clean up textures with redundant "Assets" prefix
Almost all of these textures have identical counterparts without the
"Assets" prefix, so the duplicates with the prefix are removed.
One file (AssetstldnPeriscopeViewSmlBlack.tga) has no counterpart
without the prefix, so it is renamed.
Two files (AssetstldnCrateSide01.tga and AssetstldnSkyFast.tga)
are different from their counterparts without the prefix, so they are
not changed.