Fix solid bendy bridge in Teledahn.
[PXPhysical] Bendy Bridge
Changed the member group from kGroupDynamic to kGroupStatic. The
previous setting is for kickables (which this bridge is not). PhysX does
not allow for non-kinematic dynamic bodies to have triangle mesh bounds.
Because of the improper member group, we were forcing PhysX to treat the
triangle mesh bounds as a convex hull, making the whole bridge solid.

Or, as Simulation.0.log says:
```
WARNING: 'BendyBridge' is a dynamic triangle mesh; this is not supported
in PhysX 4... forcing to convex hull, sorry.
```