Restore missing footnote in King Kerath notebook
Based on a pre-release version of the notebook that was preserved by the
Guild of Archivists (formerly DPWR), the footnotes were slightly changed
at some point during development (before the release of ABM). In
particular, the first asterisk (next to a Kerath quote) was removed from
the text, and the third footnote entry (referencing a Mee-Dis War book)
was removed from the footnotes.

It appears that these changes were accidental. The removed asterisk
doesn't match the removed footnote, and the remaining asterisks and
footnotes no longer match up very well with each other. This is most
noticeable with the "Taken from Kerath's public speech" footnote: in the
pre-release version, it matches up with a Kerath quote, but in the
release version, it corresponds to a paragraph about D'ni prophets that
doesn't contain anything said by Kerath.

This commit re-adds the previously removed asterisk and Mee-Dis War
footnote, which restores the footnotes to the more consistent
pre-release state. The French and German translations are updated
accordingly, with the translation of the footnote copied from other
notebooks. The Italian and Spanish translations are *not* updated in
this commit, because they are incomplete and don't contain a translation
of the Mee-Dis War footnote anywhere.

Source:
https://archive.guildofarchivists.org/wiki/Reference:DRC_research_notebooks/King_Kerath/2002_DPWR_transcription
https://archive.guildofarchivists.org/w/index.php?title=Reference:DRC_research_notebooks/King_Kerath&oldid=1481