Fix sentence containing percent symbol in King Hemelin notebook
One sentence in the King Hemelin notebook was partially destroyed at
some point, apparently as the result of a regex replace gone wrong.
Someone replaced "%.*s" with "%256s", probably meaning to adjust some
printf-style format strings, but this accidentally also matched a
literal percent symbol in the King Hemelin notebook and deleted some of
the following words.

In the English version, someone apparently noticed this and removed the
destroyed part of the sentence, but did not fully undo the incorrect
change. The French and German translations were never fixed.

This commit fully restores the sentence, based on the text from Uru CC,
where it was not yet broken. The English original and French and German
translations are fixed (there is no Italian or Spanish translation for
this notebook).