Replace non-Latin-1 characters in .loc files
Uru's fonts are limited to Latin-1 characters. If any characters outside
that range are used in in-game texts, the characters are either not
displayed at all or are replaced with a question mark symbol.

This change replaces all non-Latin-1 characters in all .loc files with
ASCII equivalents, so that they are displayed properly in-game.
Specifically, the following replacements are performed:
* Unicode dashes to ASCII hyphen-minus ([–—] to -)
* "Curly" quotes to "straight" quotes ([‘’‚] to ', [“”„] to ")
* Unicode ellipsis to ASCII dots (… to ...)
* œ to oe (sorry France)