Fix percent symbol in King Hemelin notebook, for real this time
The localization system *always* tries to interpret placeholders
(e. g. %1) in translations, even if it's just constant text that's not
supposed to have placeholders. To display the percent symbols here
literally, they need to be escaped (using a backslash,
counter-intuitively).

An unescaped percent symbol that doesn't form a valid placeholder will
cause strange behavior. In this case, the entire journal content before
the percent symbol was silently ignored, resulting in missing text and
broken formatting.