Consistently use guillemets in French translation
The French translation was inconsistent about its quote style. Some
parts used guillemets, but others used double quotes. As far as I know,
guillemets are the standard quote form in French.

Note that in French, there are spaces on the inside of the guillemets as
well, not just on the outside as in most other languages. These inside
spaces should be non-breaking spaces, so that the guillemets are
line-wrapped together with the text being quoted.