Use non-breaking spaces around French punctuation where appropriate
In French, some punctuation characters (question mark, exclamation mark,
colon, semicolon) also have a space before the symbol, not just after
the symbol as in most other languages. This preceding space should be a
non-breaking space, so that the punctuation is not wrapped onto a new
line separate from the preceding word.