Add spaces in numbered lists in French Great Zero notebook
This is mostly a cosmetic thing for consistency with the other
languages, but it also works around an unfortunate line wrapping
problem: the line "2.Une fois que vous disposez d'un KI version 2.0"
gets wrapped at the version number, between "." and "0". This is
probably because it interprets the "." as an end of sentence, and there
seems to be no good way to prevent this (we can't use a non-breaking
space, because we don't want a space). By adding the spaces in the
numbered lists, the problematic line becomes long enough that the entire
version number gets wrapped onto the next line, as expected.