Add linking panels and new link point into Gahreesen
grsnTrainingCntrLinkRm
Added a new subworld that rotates at the same rate as the scenery. Subworld is located around the small metal disk inside the room above the prison area.
Placed a new linking point on it to give an alternate spot to link into Gahreesen for possible easier access to the Wall.
This required several new scene objects with all the things that entails.

bkBookImage
[MipMap - xLinkPanelGarrisonPurpleControlRoom*1#0.hsm] New panel for purple wall area
[MipMap - xLinkPanelGarrisonYellowControlRoom*1#0.hsm] New panel for yellow wall area
[MipMap - xlinkpanelgarrisoncontrolroom*1#0.hsm] New panel for control center above the prison area
[Image Lib - StandardBookImageLib] add 3 new panels