Update EderDelin CubicEnvironmap for accuracy.
The environment map was previously a duplicate of the incorrect
environment map from Eder Tsogal. The end result is a very slight change
in coloration at the top of the bahro door from white to pink. The bahro
door in Eder Delin is too shrouded in shadow for this change to be
readily apparent.