Add sources for GUI_BkBookImages.
This adds a JSON file and all of the source images required to rebuild
GUI_District_BkBookImages.prp using my MakeImageLibPRP.py script. The
source images were all ripped from the previous BkBookImages.prp and
compared with the files already in the sources directory. If an
identical lossless version already existed, the lossy version from the
PRP was discarded. Otherwise, the lossy version from the PRP has been
added here and is re-imported to the the new PRP.

The linking panels are primarily the trouble spots here - they were
mostly reshot for MOUL, but what we have here is mostly the TPotS
linking panel images. Some of them are identical. Many of the Yeesha
journey Ages use the same angle but have been brightened somewhat. As
stated above, I brought in the MOUL reshots if there was any discernible
difference. No effort was made to correct problems with any linking
panels. So, for example, the lower pellet cave linking panel still has a
cursor in it.

The resuling diff on BkBookImages.prp should be minimal. There were some
RLE changes reevaluated, and any PNGs in BkBookImages.prp were
compressed to JPEG. In my work, I discovered using lossless compression
resulted in a 60 MB BkBookImages.prp, which is undesirable.